Added more control over links and support for 3 apps
- You can now give a level to each app. 0 means 'no tweet', 1 means
'tweet with default app link' and 2 means 'tweet with original app
link'.
- A default_app setting has been added. It lets you decide which app is
the default one, previously Steemit was forced on you. This setting must
be filled by any app that supports reading posts from any other app. A
list will soon be added to the README file. If this setting is invalid, the default app will be Steemit.
- Added support for Insteem, Steemd and SteemDB.